+++ bgplot/__init__.py
@@ -1,2 +1,3 @@
 from .features.graphics.graphics import Graphics
 from .features import operations as op
+from .core.logger import Logger