+++ bgplot/entities/point.py
@@ -2,6 +2,25 @@ from typing import NamedTuple
 
 
 class Point(NamedTuple):
+    """
+    a named tuple that contains the information of a 3D point in space
+
+    Attributes
+    ----------
+    x : float
+        the x value of the point
+
+    y : float
+        the y value of the point
+
+    z : float
+        the z value of the point
+
+    Notes
+    -----
+    supports addition and subtraction with other Point class instances.
+    Supports scalar multiplication
+    """
     x: float
     y: float
     z: float