+++ bgplot/entities/vector.py
@@ -3,6 +3,25 @@ from typing import NamedTuple
 
 
 class Vector(NamedTuple):
+    """
+    a named tuple that contains the information of a 3D free vector in space
+
+    Attributes
+    ----------
+    u : float
+        the value of the first element on the u-v-w axes
+
+    v : float
+        the value of the second element on the u-v-w axes
+
+    w : float
+        the value of the third element on the u-v-w axes
+
+    Notes
+    -----
+    supports addition and subtraction with other Vector class instances.
+    supports scalar multiplication
+    """
     u: float
     v: float
     w: float
@@ -17,6 +36,14 @@ class Vector(NamedTuple):
         return Vector(self.u-other.u, self.v-other.v, self.w-other.w)
 
     def normalize(self):
+        """
+        normalizes the vector so its magnitude is unitary
+
+        Returns
+        -------
+        out : ~.entities.vector.Vector
+            the normalized instance of the vector
+        """
         length: float = np.linalg.norm(self)
         u: float = self.u / length
         v: float = self.v / length